+++ src/lazyload-image.directive.ts
@@ -1,8 +1,5 @@
-import 'rxjs/add/operator/let';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/debounceTime';
-import { Observable } from 'rxjs/Observable';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { Observable, ReplaySubject } from 'rxjs';
+import { switchMap, debounceTime, startWith } from 'rxjs/operators';
 import {
     AfterContentInit,
     Directive,
@@ -17,16 +14,16 @@ import {
 } from '@angular/core';
 import { getScrollListener } from './scroll-listener';
 import { lazyLoadImage } from './lazyload-image';
-
-const target = typeof window !== 'undefined' ? window : undefined;
+import { isWindowDefined } from './utils';
 
 interface LazyLoadImageDirectiveProps {
     lazyImage: string;
     defaultImage: string;
     errorImage: string;
-    scrollTarget: Object;
+    scrollTarget: any;
     scrollObservable: Observable<Event>;
     offset: number;
+    useSrcset: boolean;
 }
 
 export interface IOnLoadedPayload {
@@ -41,14 +38,16 @@ export class LazyLoadImageDirective implements OnChanges, AfterContentInit, OnDe
     @Input('lazyLoad') lazyImage;   // The image to be lazy loaded
     @Input() defaultImage: string;  // The image to be displayed before lazyImage is loaded
     @Input() errorImage: string;    // The image to be displayed if lazyImage load fails
-    @Input() scrollTarget = target; // Change the node we should listen for scroll events on, default is window
+    @Input() scrollTarget: any;     // Scroll container that contains the image and emits scoll events
     @Input() scrollObservable;      // Pass your own scroll emitter
     @Input() offset: number;        // The number of px a image should be loaded before it is in view port
+
+    @Input() useSrcset: boolean;    // Whether srcset attribute should be used instead of src    
     @Output() onLoad: EventEmitter<IOnLoadedPayload> = new EventEmitter(); // Callback when an image is loaded
+
     private propertyChanges$: ReplaySubject<LazyLoadImageDirectiveProps>;
     private elementRef: ElementRef;
     private ngZone: NgZone;
-    private platformId: string;
     private scrollSubscription;
 
     constructor(el: ElementRef, ngZone: NgZone) {
@@ -64,42 +63,42 @@ export class LazyLoadImageDirective implements OnChanges, AfterContentInit, OnDe
             errorImage: this.errorImage,
             scrollTarget: this.scrollTarget,
             scrollObservable: this.scrollObservable,
-            offset: this.offset,
+            offset: this.offset | 0,
+            useSrcset: this.useSrcset
         });
     }
 
     ngAfterContentInit() {
-        /**
-         * Disable lazy load image in server side
-         * @see https://github.com/tjoskar/ng-lazyload-image/issues/178
-         * @see https://github.com/tjoskar/ng-lazyload-image/issues/183
-         */
-        if (typeof window === 'undefined') {
+        // Disable lazy load image in server side
+        if (!isWindowDefined()) {
             return null;
         }
 
         this.ngZone.runOutsideAngular(() => {
             let scrollObservable: Observable<Event>;
             if (this.scrollObservable) {
-                scrollObservable = this.scrollObservable.startWith('');
+                scrollObservable = this.scrollObservable.pipe(startWith(''));
             } else {
-                scrollObservable = getScrollListener(this.scrollTarget);
+                const windowTarget = isWindowDefined() ? window : undefined;
+                scrollObservable = getScrollListener(this.scrollTarget || windowTarget);
             }
-            this.scrollSubscription = this.propertyChanges$
-                .debounceTime(10)
-                .switchMap(props => scrollObservable.let(
+            this.scrollSubscription = this.propertyChanges$.pipe(
+                debounceTime(10),
+                switchMap(props => scrollObservable.pipe(
                     lazyLoadImage(
                         this.elementRef.nativeElement,
                         props.lazyImage,
                         props.defaultImage,
                         props.errorImage,
-                        props.offset
+                        props.offset,
+                        props.useSrcset,
+                        props.scrollTarget
                     )
                 ))
-                .subscribe(success => this.onLoad.emit({
-                    success: success,
-                    element: this.elementRef
-                }));
+            ).subscribe(success => this.onLoad.emit({
+                success: success,
+                element: this.elementRef
+            }));
         });
     }